+++ Tests/MongoDb.TestApplication/MongoDb.TestApplication.Api/Configuration/SwashbuckleConfiguration.cs
@@ -35,10 +35,7 @@ namespace MongoDb.TestApplication.Api.Configuration
 
         public static void UseSwashbuckle(this IApplicationBuilder app)
         {
-            app.UseSwagger(
-                options =>
-                {
-                });
+            app.UseSwagger();
             app.UseSwaggerUI(
                 options =>
                 {