+++ Tests/Publish.CleanArch.MassTransit.OutboxNone.TestApplication/Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Api/Configuration/SwashbuckleConfiguration.cs
@@ -58,10 +58,7 @@ namespace Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Api.Configura
 
         public static void UseSwashbuckle(this IApplicationBuilder app)
         {
-            app.UseSwagger(
-                options =>
-                {
-                });
+            app.UseSwagger();
             app.UseSwaggerUI(
                 options =>
                 {